pyproject.toml
[tool.poetry]
name = "modelguard"
version = "0.1.0"
description = ""
authors = ["isheng <z26488200@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "0.115.0"
uvicorn = {version = "0.30.6", extras = ["standard"]}
minio = "7.2.8"
python-multipart = "0.0.9"
pydantic = "2.9.2"
scikit-learn = "1.5.2"
numpy = "1.26.4"
python-jose = {version = "3.3.0", extras = ["cryptography"]}
bcrypt = "4.2.1"
httpx = "0.27.2"
aiofiles = "24.1.0"
python-dotenv = "1.0.1"

[tool.poetry.group.dev.dependencies]
pytest = "*"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
